requirements.txt
streamlit
torch
transformers
scikit-learn
scikit-surprise
pandas
numpy==1.26.4  
nrclex
datasets